Move EncryptCommon class to it's own file
