+++ COSE/COSE.Net45.csproj
@@ -57,6 +57,7 @@
     <Compile Include="Attributes.cs" />
     <Compile Include="ChaCha20Poly1305.cs" />
     <Compile Include="Encrypt0Mesage.cs" />
+    <Compile Include="EncryptCommon.cs" />
     <Compile Include="EncryptMessage.cs" />
     <Compile Include="Key.cs" />
     <Compile Include="KeySet.cs" />
