+++ COSE/Encrypt0Mesage.cs
@@ -6,27 +6,40 @@ namespace Com.AugustCellars.COSE
 {
     public class Encrypt0Message : EncryptCommon
     {
+        /// <summary>
+        /// Implement the COSE_Encrypt0 protocol element from RFC 8215.
+        /// This will emit the tag and the content
+        /// </summary>
         public Encrypt0Message() : base(true, true)
         {
             _context = "Encrypt0";
             m_tag = Tags.Encrypt0;
         }
 
+        /// <summary>
+        /// Implement the COSE_Encrypt0 protocol element from RFC 8215.
+        /// </summary>
+        /// <param name="fEmitTag">emit leading tag</param>
+        /// <param name="fEmitContent">emit message content</param>
         public Encrypt0Message(bool fEmitTag, bool fEmitContent = true) : base(fEmitTag, fEmitContent)
         {
             _context = "Encrypt0";
             m_tag = Tags.Encrypt0;
         }
 
-        public virtual void DecodeFromCBORObject(CBORObject obj)
+        /// <summary>
+        /// Given a CBOR tree, try and parse the tree into an Encrypt0 item.
+        /// </summary>
+        /// <param name="cbor"></param>
+        public virtual void DecodeFromCBORObject(CBORObject cbor)
         {
-            if (obj.Count != 3) throw new CoseException("Invalid Encrypt0 structure");
+            if (cbor.Count != 3) throw new CoseException("Invalid Encrypt0 structure");
 
             //  Protected values.
-            if (obj[0].Type == CBORType.ByteString)
+            if (cbor[0].Type == CBORType.ByteString)
             {
-                if (obj[0].GetByteString().Length == 0) objProtected = CBORObject.NewMap();
-                else objProtected = CBORObject.DecodeFromBytes(obj[0].GetByteString());
+                if (cbor[0].GetByteString().Length == 0) objProtected = CBORObject.NewMap();
+                else objProtected = CBORObject.DecodeFromBytes(cbor[0].GetByteString());
                 if (objProtected.Type != CBORType.Map) throw new CoseException("Invalid Encrypt0 structure");
             }
             else
@@ -35,20 +48,25 @@ namespace Com.AugustCellars.COSE
             }
 
             //  Unprotected attributes
-            if (obj[1].Type == CBORType.Map) objUnprotected = obj[1];
+            if (cbor[1].Type == CBORType.Map) objUnprotected = cbor[1];
             else throw new CoseException("Invalid Encrypt0 structure");
 
             // Cipher Text
-            if (obj[2].Type == CBORType.ByteString) _rgbEncrypted = obj[2].GetByteString();
-            else if (!obj[2].IsNull)
+            if (cbor[2].Type == CBORType.ByteString) _rgbEncrypted = cbor[2].GetByteString();
+            else if (!cbor[2].IsNull)
             {               // Detached content - will need to get externally
                 throw new CoseException("Invalid Encrypt0 structure");
             }
         }
 
+        /// <summary>
+        /// Encode the COSE Encrypt0 item to a CBOR tree.
+        /// <see cref="Encrypt"/> must be done prior to calling this function.
+        /// </summary>
+        /// <returns></returns>
         public override CBORObject Encode()
         {
-            CBORObject obj;
+            CBORObject cbor;
 
             if (_rgbEncrypted == null) throw new CoseException("Must call Encrypt first");
 
@@ -62,27 +80,36 @@ namespace Com.AugustCellars.COSE
                     }
                 }
             }
-            obj = CBORObject.NewArray();
+            cbor = CBORObject.NewArray();
 
             if (objProtected.Count > 0) {
-                obj.Add(objProtected.EncodeToBytes());
+                cbor.Add(objProtected.EncodeToBytes());
             }
-            else obj.Add(CBORObject.FromObject(new byte[0]));
+            else cbor.Add(CBORObject.FromObject(new byte[0]));
 
-            obj.Add(objUnprotected); // Add unprotected attributes
+            cbor.Add(objUnprotected); // Add unprotected attributes
 
-            if (m_emitContent) obj.Add(_rgbEncrypted);      // Add ciphertext
-            else obj.Add(CBORObject.Null);
+            if (m_emitContent) cbor.Add(_rgbEncrypted);      // Add ciphertext
+            else cbor.Add(CBORObject.Null);
 
-            return obj;
+            return cbor;
         }
 
+        /// <summary>
+        /// Attempt to decrypt the message.
+        /// </summary>
+        /// <param name="rgbKey">key to be used for decryption</param>
+        /// <returns>decrypted content</returns>
         public byte[] Decrypt(byte[] rgbKey)
         {
             DecryptWithKey(rgbKey);
             return rgbContent;
         }
 
+        /// <summary>
+        /// Encrypt the message with the provided key
+        /// </summary>
+        /// <param name="rgbKey">key for encryption</param>
         public void Encrypt(byte[] rgbKey)
         {
             EncryptWithKey(rgbKey);
